Demo 2502 add mysql integration v2 (#19)
* [DEMO-2502] Add mysql on-host newrelic integration

* Added the tasks & template for installing the integration, as well as
verifying that it works once installed. The integration installation
uses the public new relic infra ansible role.
* The role is 'onafterstart' because this needs mysql to be installed.
The installation of mysql occurs in 'start', so this must follow that.

* [DEMO-2502] Make port optional

* Added a new port variable which is effectively optional. It uses the
pre-existing variable if its defined, otherwise defaults to 3306.

* [DEMO-2502] Generate newrelic user password

* Remove the sharing of passwords and instead generate the password as
part of the play. The password generation is idempotent, so multiple
deployments use the first password generated.

* [DEMO-2502] Remove existing labels, add tags as labels

* Removed existing labels
* Added a step to the play to add tags as labels

* [DEMO-2502] Swap out hard wait with loop until

* Instead of waiting for a specific amount of time, use loop until.

* [DEMo-2502] Add SELECT permission to newrelic integration user

* Adding SELECT permission to the newrelic integration user based on PR
comment that this is needed for certain setups.

* [DEMO-2502] Move clear logs step

* Moving the step that clears logs after the agent role. This allows us
to remove the stat step, since the log file will exist.

* [DEMO-2502] Improvements in organization of ansible plays

* Now the infra-agent and mysql integration are installed separately.
* The mysql integration is installed in the newly added instrument step
(previously was in the onafterstartstep).
* Password generation still happens during the play, but the usage of
the password is now simplified. On each deployment, the password of the
user is updated. This allowed me to remove the other steps.
* Added log_file configuration to the newrelic infrastructure agent to
be able to look at its logs for verifying things are working.

* [DEMO-2502] Renamed folder structure

* Instead of 'databases/linux/mysql/roles' we now have
'databases/mysql/linux/roles'. This just elevates the database type in
the hierarchy. Ideally, we would be able to instrument the DB on any OS,
so the OS folder would be irrelevant, but thats far in the future.

* [DEMO-2502] Rename password variable

* Rename root_user_password to database_root_password for consistency
with other fields in the deployment config.

* [DEMO-2502] Add comment explaining loop

* Added a comment explaining the loop at the top of installing the
instrumentation play.
* Additionally, moved the check to the top of the play since its
essentially a pre-requisite.

* [DEMO-2502] Remove port default value